+++ app/serializers/forest_liana/mixpanel_event_serializer.rb
@@ -1,0 +1,24 @@
+module ForestLiana
+  class MixpanelEventSerializer
+    include JSONAPI::Serializer
+
+    attribute :id
+    attribute :event
+    attribute :city
+    attribute :region
+    attribute :timezone
+    attribute :os
+    attribute :osVersion
+    attribute :country
+    attribute :date
+    attribute :browser
+
+    def self_link
+      nil
+    end
+
+    def type
+      @options[:context][:type] || 'mixpanel_events'
+    end
+  end
+end